poetry: Remove black from dev-dependencies
Until now, we had to define `black` version in multiple locations
(`pyproject.toml`, `pre-commit-config.yaml` etc) and keep them in sync.
In order to avoid this, let's use `black` from `pre-commit` and remove it
from `pyproject.toml`.

For the record, as long as `black` was in `beta`, using the auto-updating
feature of `pre-commit` was prone to breaking changes. That's why we were
also defining black on pyproject.toml` too. Now that black is considered
 stable, I don't think this is such a problem anymore.