pre-commit: Disable `fail_fast`
Whenever dependencies change, `poetry-export` and `poetry-export-dev`
are failing. With `fail_fast = true` in order to generate both
`requirements.txt` and `requirements-dev.txt` we had to run pre-commit
twice.

With `fail_fast = false` both files get generated in a single run. Since
the other checks are relatively fast, this should not be a big issue.