prospector: Disable `pylint`.
This is a test. I am not sure if we want to completely ditch `pylint` or
not.  The problem is that `pylint` is quite slow. By ditching it, `make
lint` returns almost instantly. Moreover `pycodestyle` covers most of
the checks.

Note 1: I am leaving the configuration in place, so that we can easily
re-enable it in the future.

Note 2: I removed the `pylint: disable` comments to improve the
readability of the code. Adding them back is fairly trivial so it should
not be a problem.