coops: Remove typepigeon dependency.
`typepigeon` changed its API in the recently released version 2. We could add some
compatibility layer to support both v1 and v2 APIs, but we only use it for  a very simple thing,
Therefore removing it feels better for maintenance in the long run.

The changes could probably be quite a bit simpler, but I tried to do exactly the same thing
that typepigeon does, just in case anyone uses searvey in a non-standard way. E.g. by passing
`'h'` instead of `COOPS_Interval.H` etc.

The rest are mypy fixes. Again not very pretty but typing needs quite a bit of work in `coops.py`
anyhow.