Resolve dates dynamically
Due to early-binding, if no value was passed for `endtime` then we were always getting the
same date (i.e. the date when the code was imported). With this commit we replace the
default value with a sentinel value (`TODAY`) and we simplify the IOC/USGS functions by
hiding the `endpoint` logic into `resolve_date()`.

Fixes #66